moved recuder file to context directory. Added selecteam name to state
