+++ frontend/src/components/SelectedTeam.js
@@ -1,7 +1,9 @@
-import React from 'react';
+/* eslint-disable no-empty-pattern */
+import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
+import { useStateValue } from '../contexts/state';
 import NotFound from './NotFound';
-import { useHistory } from 'react-router-dom';
+import { useHistory, useParams } from 'react-router-dom';
 import theme from '../styles/theme';
 import BreadcrumbNav from './BreadcrumbNav';
 import { pickIcon } from './TestIcon';
@@ -12,6 +14,13 @@ import { css, jsx } from '@emotion/core';
 
 const SelectedTeam = ({ selectedTeam }) => {
     const [t] = useTranslation(['team']);
+
+    const { name } = useParams();
+    const [{}, dispatch] = useStateValue();
+
+    useEffect(() => {
+        dispatch({ type: 'setSelectedTeam', selectedTeam: name });
+    }, [dispatch, name]);
 
     const cardStyles = css`
         background-color: #fafafa;
