+++ frontend/src/contexts/state.js
@@ -23,7 +23,8 @@ const initialState = {
     parentData: {
         seriesData: null,
         buildData: null
-    }
+    },
+    selectedTeam: ''
 };
 
 export const StateProvider = ({ reducer, children }) => {
